Add Play functionality to CallTree
Implemented the Play case class in CallTree to support audio play functionality in Dialogue State. The Play class stores a URL to an audio file to be played. Also updated corresponding methods in Twilio, TexML, and Telnyx services to handle this new case. This will enhance the versatility of the system by supporting automated playback of audio files in addition to text-to-speech.