Add title parameter to Menu methods for CallTree
Updated the Menu methods in CallTree in scala application to accept a title parameter. This change will allow the addition of a 'title' to the menu prompt, rather than having to prepend it to the menu. This way the user can input a response during the title, even before hearing the choices