Added 'controls' attribute to audio element
This commit includes the addition of 'controls' attribute to HTML audio elements within `toHtml` methods in `twilio/Twiml.scala` and `telnyx/TeXML.scala`. This attribute enables built-in audio controls in the browser. Additionally, this update includes the declaration of `controls` in `^` objects in `telnyx/TagsBundle.scala` and `twilio/TagsBundle.scala`.